Extract desktop state thread helpers
