+++ src/composables/useDesktopState.ts
@@ -94,840 +94,71 @@ import {
   writeSelectedCollaborationModeForContext,
 } from './desktopStateStorage'
 export { isThreadUnreadByLastRead } from './desktopStateStorage'
-
-function flattenThreads(groups: UiProjectGroup[]): UiThread[] {
-  return groups.flatMap((group) => group.threads)
-}
-
-export function findAdjacentThreadId(threads: UiThread[], threadId: string): string {
-  const targetIndex = threads.findIndex((thread) => thread.id === threadId)
-  if (targetIndex < 0) return ''
-  return threads[targetIndex + 1]?.id ?? threads[targetIndex - 1]?.id ?? ''
-}
-
-const EVENT_SYNC_DEBOUNCE_MS = 220
-const BACKGROUND_THREAD_PAGINATION_DELAY_MS = 10_000
-const RATE_LIMIT_REFRESH_DEBOUNCE_MS = 500
-const TURN_START_FOLLOW_UP_SYNC_DELAY_MS = 3000
-const RECENT_THREAD_MESSAGE_LOAD_REUSE_MS = 2000
-const REASONING_EFFORT_OPTIONS: ReasoningEffort[] = ['none', 'minimal', 'low', 'medium', 'high', 'xhigh']
-const GLOBAL_SERVER_REQUEST_SCOPE = '__global__'
-const MODEL_FALLBACK_ID = 'gpt-5.4-mini'
-
-function mergeProjectOrder(previousOrder: string[], incomingGroups: UiProjectGroup[]): string[] {
-  const nextOrder: string[] = []
-
-  for (const projectName of previousOrder) {
-    if (!nextOrder.includes(projectName)) {
-      nextOrder.push(projectName)
-    }
-  }
-
-  for (const group of incomingGroups) {
-    if (!nextOrder.includes(group.projectName)) {
-      nextOrder.push(group.projectName)
-    }
-  }
-
-  return areStringArraysEqual(previousOrder, nextOrder) ? previousOrder : nextOrder
-}
-
-function orderGroupsByProjectOrder(incoming: UiProjectGroup[], projectOrder: string[]): UiProjectGroup[] {
-  const incomingByName = new Map(incoming.map((group) => [group.projectName, group]))
-  const ordered: UiProjectGroup[] = projectOrder
-    .map((projectName) => incomingByName.get(projectName) ?? null)
-    .filter((group): group is UiProjectGroup => group !== null)
-
-  for (const group of incoming) {
-    if (!projectOrder.includes(group.projectName)) {
-      ordered.push(group)
-    }
-  }
-
-  return ordered
-}
-
-function areStringArraysEqual(first?: string[], second?: string[]): boolean {
-  const left = Array.isArray(first) ? first : []
-  const right = Array.isArray(second) ? second : []
-  if (left.length !== right.length) return false
-  for (let index = 0; index < left.length; index += 1) {
-    if (left[index] !== right[index]) return false
-  }
-  return true
-}
-
-function reorderStringArray(items: string[], fromIndex: number, toIndex: number): string[] {
-  if (fromIndex < 0 || fromIndex >= items.length || toIndex < 0 || toIndex >= items.length) {
-    return items
-  }
-
-  if (fromIndex === toIndex) {
-    return items
-  }
-
-  const next = [...items]
-  const [moved] = next.splice(fromIndex, 1)
-  next.splice(toIndex, 0, moved)
-  return next
-}
-
-function areCommandExecutionsEqual(first?: CommandExecutionData, second?: CommandExecutionData): boolean {
-  if (!first && !second) return true
-  if (!first || !second) return false
-  return first.status === second.status && first.aggregatedOutput === second.aggregatedOutput && first.exitCode === second.exitCode
-}
-
-function arePlanStepsEqual(first: UiPlanStep[] = [], second: UiPlanStep[] = []): boolean {
-  if (first.length !== second.length) return false
-  for (let index = 0; index < first.length; index += 1) {
-    if (first[index]?.step !== second[index]?.step || first[index]?.status !== second[index]?.status) {
-      return false
-    }
-  }
-  return true
-}
-
-function arePlanDataEqual(first?: UiPlanData, second?: UiPlanData): boolean {
-  if (!first && !second) return true
-  if (!first || !second) return false
-  return (
-    first.explanation === second.explanation &&
-    first.isStreaming === second.isStreaming &&
-    arePlanStepsEqual(first.steps, second.steps)
-  )
-}
-
-function isUnsupportedChatGptModelError(error: unknown): boolean {
-  if (!(error instanceof Error)) return false
-  const message = error.message.toLowerCase()
-  return (
-    message.includes('not supported when using codex with a chatgpt account') ||
-    message.includes('model is not supported') ||
-    message.includes('requires a newer version of codex')
-  )
-}
-
-function areMessageFieldsEqual(first: UiMessage, second: UiMessage): boolean {
-  return (
-    first.id === second.id &&
-    first.role === second.role &&
-    first.text === second.text &&
-    areStringArraysEqual(first.images, second.images) &&
-    areUiFileChangesEqual(first.fileChanges, second.fileChanges) &&
-    first.fileChangeStatus === second.fileChangeStatus &&
-    first.messageType === second.messageType &&
-    first.rawPayload === second.rawPayload &&
-    first.isUnhandled === second.isUnhandled &&
-    areCommandExecutionsEqual(first.commandExecution, second.commandExecution) &&
-    arePlanDataEqual(first.plan, second.plan) &&
-    first.turnId === second.turnId &&
-    first.turnIndex === second.turnIndex &&
-    first.isAutomationRun === second.isAutomationRun &&
-    first.automationDisplayName === second.automationDisplayName
-  )
-}
-
-function areMessageArraysEqual(first: UiMessage[], second: UiMessage[]): boolean {
-  if (first.length !== second.length) return false
-  for (let index = 0; index < first.length; index += 1) {
-    if (first[index] !== second[index]) return false
-  }
-  return true
-}
-
-function mergeMessages(
-  previous: UiMessage[],
-  incoming: UiMessage[],
-  options: { preserveMissing?: boolean } = {},
-): UiMessage[] {
-  const previousById = new Map(previous.map((message) => [message.id, message]))
-  const incomingById = new Map(incoming.map((message) => [message.id, message]))
-
-  const mergedIncoming = incoming.map((incomingMessage) => {
-    const previousMessage = previousById.get(incomingMessage.id)
-    if (previousMessage && areMessageFieldsEqual(previousMessage, incomingMessage)) {
-      return previousMessage
-    }
-    return incomingMessage
-  })
-
-  if (options.preserveMissing !== true) {
-    return areMessageArraysEqual(previous, mergedIncoming) ? previous : mergedIncoming
-  }
-
-  const mergedFromPrevious = previous.map((previousMessage) => {
-    const nextMessage = incomingById.get(previousMessage.id)
-    if (!nextMessage) {
-      return previousMessage
-    }
-    if (areMessageFieldsEqual(previousMessage, nextMessage)) {
-      return previousMessage
-    }
-    return nextMessage
-  })
-
-  const previousIdSet = new Set(previous.map((message) => message.id))
-  const appended = mergedIncoming.filter((message) => !previousIdSet.has(message.id))
-  const merged = [...mergedFromPrevious, ...appended]
-
-  return areMessageArraysEqual(previous, merged) ? previous : merged
-}
-
-function areUiFileChangesEqual(first?: UiFileChange[], second?: UiFileChange[]): boolean {
-  if (!first && !second) return true
-  if (!first || !second) return false
-  if (first.length !== second.length) return false
-  for (let index = 0; index < first.length; index += 1) {
-    const firstChange = first[index]
-    const secondChange = second[index]
-    if (
-      firstChange.path !== secondChange.path ||
-      firstChange.operation !== secondChange.operation ||
-      firstChange.movedToPath !== secondChange.movedToPath ||
-      firstChange.diff !== secondChange.diff ||
-      firstChange.addedLineCount !== secondChange.addedLineCount ||
-      firstChange.removedLineCount !== secondChange.removedLineCount
-    ) {
-      return false
-    }
-  }
-  return true
-}
-
-function normalizeMessageText(value: string): string {
-  return value.replace(/\s+/gu, ' ').trim()
-}
-
-function removeRedundantLiveAgentMessages(previous: UiMessage[], incoming: UiMessage[]): UiMessage[] {
-  const incomingMessageIds = new Set(incoming.map((message) => message.id))
-  const incomingAssistantTexts = new Set(
-    incoming
-      .filter((message) => message.role === 'assistant')
-      .map((message) => normalizeMessageText(message.text))
-      .filter((text) => text.length > 0),
-  )
-
-  if (incomingAssistantTexts.size === 0) {
-    return previous
-  }
-
-  const next = previous.filter((message) => {
-    if (message.messageType !== 'agentMessage.live') return true
-    if (incomingMessageIds.has(message.id)) return false
-    const normalized = normalizeMessageText(message.text)
-    if (normalized.length === 0) return false
-    return !incomingAssistantTexts.has(normalized)
-  })
-
-  return next.length === previous.length ? previous : next
-}
-
-function removePersistedLiveMessages(previous: UiMessage[], incoming: UiMessage[]): UiMessage[] {
-  const incomingIds = new Set(incoming.map((message) => message.id))
-  const next = previous.filter((message) => !incomingIds.has(message.id))
-  return next.length === previous.length ? previous : next
-}
-
-function upsertMessage(previous: UiMessage[], nextMessage: UiMessage): UiMessage[] {
-  const existingIndex = previous.findIndex((message) => message.id === nextMessage.id)
-  if (existingIndex < 0) {
-    return [...previous, nextMessage]
-  }
-
-  const existing = previous[existingIndex]
-  if (areMessageFieldsEqual(existing, nextMessage)) {
-    return previous
-  }
-
-  const next = [...previous]
-  next.splice(existingIndex, 1, nextMessage)
-  return next
-}
-
-type TurnSummaryState = {
-  turnId: string
-  durationMs: number
-}
-
-type TurnActivityState = {
-  label: string
-  details: string[]
-}
-
-type TurnErrorState = {
-  message: string
-  transient: boolean
-}
-
-type TurnStartedInfo = {
-  threadId: string
-  turnId: string
-  startedAtMs: number
-}
-
-type TurnCompletedInfo = {
-  threadId: string
-  turnId: string
-  completedAtMs: number
-  startedAtMs?: number
-}
-
-const WORKED_MESSAGE_TYPE = 'worked'
-
-function parseIsoTimestamp(value: string): number | null {
-  if (!value) return null
-  const ms = new Date(value).getTime()
-  return Number.isNaN(ms) ? null : ms
-}
-
-function formatTurnDuration(durationMs: number): string {
-  if (!Number.isFinite(durationMs) || durationMs <= 0) {
-    return '<1s'
-  }
-
-  const totalSeconds = Math.max(1, Math.round(durationMs / 1000))
-  const hours = Math.floor(totalSeconds / 3600)
-  const minutes = Math.floor((totalSeconds % 3600) / 60)
-  const seconds = totalSeconds % 60
-  const parts: string[] = []
-
-  if (hours > 0) {
-    parts.push(`${hours}h`)
-  }
-
-  if (minutes > 0 || hours > 0) {
-    parts.push(`${minutes}m`)
-  }
-
-  const displaySeconds = seconds > 0 || parts.length === 0 ? seconds : 0
-  parts.push(`${displaySeconds}s`)
-  return parts.join(' ')
-}
-
-function areTurnSummariesEqual(first?: TurnSummaryState, second?: TurnSummaryState): boolean {
-  if (!first && !second) return true
-  if (!first || !second) return false
-  return first.turnId === second.turnId && first.durationMs === second.durationMs
-}
-
-function areTurnActivitiesEqual(first?: TurnActivityState, second?: TurnActivityState): boolean {
-  if (!first && !second) return true
-  if (!first || !second) return false
-  if (first.label !== second.label) return false
-  if (first.details.length !== second.details.length) return false
-  for (let index = 0; index < first.details.length; index += 1) {
-    if (first.details[index] !== second.details[index]) return false
-  }
-  return true
-}
-
-function buildTurnSummaryMessage(summary: TurnSummaryState): UiMessage {
-  return {
-    id: `turn-summary:${summary.turnId}`,
-    role: 'system',
-    text: `Worked for ${formatTurnDuration(summary.durationMs)}`,
-    messageType: WORKED_MESSAGE_TYPE,
-    turnId: summary.turnId,
-  }
-}
-
-function findLastAssistantMessageIndex(messages: UiMessage[]): number {
-  for (let index = messages.length - 1; index >= 0; index -= 1) {
-    if (messages[index].role === 'assistant') {
-      return index
-    }
-  }
-  return -1
-}
-
-function insertTurnSummaryMessage(messages: UiMessage[], summary: TurnSummaryState): UiMessage[] {
-  const summaryMessage = buildTurnSummaryMessage(summary)
-  const sanitizedMessages = messages.filter((message) => message.messageType !== WORKED_MESSAGE_TYPE)
-  const insertIndex = findLastAssistantMessageIndex(sanitizedMessages)
-  if (insertIndex < 0) {
-    return [...sanitizedMessages, summaryMessage]
-  }
-  const next = [...sanitizedMessages]
-  next.splice(insertIndex, 0, summaryMessage)
-  return next
-}
-
-function omitKey<TValue>(record: Record<string, TValue>, key: string): Record<string, TValue> {
-  if (!(key in record)) return record
-  const next = { ...record }
-  delete next[key]
-  return next
-}
-
-function omitKeys<TValue>(record: Record<string, TValue>, keys: Set<string>): Record<string, TValue> {
-  if (keys.size === 0) return record
-  let changed = false
-  const next: Record<string, TValue> = {}
-  for (const [key, value] of Object.entries(record)) {
-    if (keys.has(key)) {
-      changed = true
-      continue
-    }
-    next[key] = value
-  }
-  return changed ? next : record
-}
-
-function areThreadFieldsEqual(first: UiThread, second: UiThread): boolean {
-  return (
-    first.id === second.id &&
-    first.title === second.title &&
-    first.projectName === second.projectName &&
-    first.cwd === second.cwd &&
-    first.createdAtIso === second.createdAtIso &&
-    first.updatedAtIso === second.updatedAtIso &&
-    first.preview === second.preview &&
-    first.unread === second.unread &&
-    first.inProgress === second.inProgress &&
-    first.pendingRequestState === second.pendingRequestState
-  )
-}
-
-function areThreadArraysEqual(first: UiThread[], second: UiThread[]): boolean {
-  if (first.length !== second.length) return false
-  for (let index = 0; index < first.length; index += 1) {
-    if (first[index] !== second[index]) return false
-  }
-  return true
-}
-
-function areGroupArraysEqual(first: UiProjectGroup[], second: UiProjectGroup[]): boolean {
-  if (first.length !== second.length) return false
-  for (let index = 0; index < first.length; index += 1) {
-    if (first[index] !== second[index]) return false
-  }
-  return true
-}
-
-function pruneThreadStateMap<T>(stateMap: Record<string, T>, threadIds: Set<string>): Record<string, T> {
-  const nextEntries = Object.entries(stateMap).filter(([threadId]) => threadIds.has(threadId))
-  if (nextEntries.length === Object.keys(stateMap).length) {
-    return stateMap
-  }
-  return Object.fromEntries(nextEntries) as Record<string, T>
-}
-
-export function removeThreadFromGroups(groups: UiProjectGroup[], threadId: string): UiProjectGroup[] {
-  const normalizedThreadId = threadId.trim()
-  if (!normalizedThreadId) return groups
-
-  let changed = false
-  const nextGroups: UiProjectGroup[] = []
-
-  for (const group of groups) {
-    const nextThreads = group.threads.filter((thread) => thread.id !== normalizedThreadId)
-    const removedFromGroup = nextThreads.length !== group.threads.length
-    if (removedFromGroup) {
-      changed = true
-    }
-    if (nextThreads.length > 0) {
-      nextGroups.push(removedFromGroup ? { ...group, threads: nextThreads } : group)
-    } else if (group.threads.length === 0) {
-      nextGroups.push(group)
-    }
-  }
-
-  return changed ? nextGroups : groups
-}
-
-function mergeThreadGroups(
-  previous: UiProjectGroup[],
-  incoming: UiProjectGroup[],
-): UiProjectGroup[] {
-  const previousGroupsByName = new Map(previous.map((group) => [group.projectName, group]))
-  const mergedGroups: UiProjectGroup[] = incoming.map((incomingGroup) => {
-    const previousGroup = previousGroupsByName.get(incomingGroup.projectName)
-    const previousThreadsById = new Map(previousGroup?.threads.map((thread) => [thread.id, thread]) ?? [])
-
-    const mergedThreads = incomingGroup.threads.map((incomingThread) => {
-      const previousThread = previousThreadsById.get(incomingThread.id)
-      if (previousThread && areThreadFieldsEqual(previousThread, incomingThread)) {
-        return previousThread
-      }
-      return incomingThread
-    })
-
-    if (
-      previousGroup &&
-      previousGroup.projectName === incomingGroup.projectName &&
-      areThreadArraysEqual(previousGroup.threads, mergedThreads)
-    ) {
-      return previousGroup
-    }
-
-    return {
-      projectName: incomingGroup.projectName,
-      threads: mergedThreads,
-    }
-  })
-
-  return areGroupArraysEqual(previous, mergedGroups) ? previous : mergedGroups
-}
-
-function mergeIncomingWithLocalInProgressThreads(
-  previous: UiProjectGroup[],
-  incoming: UiProjectGroup[],
-  inProgressById: Record<string, boolean>,
-): UiProjectGroup[] {
-  const incomingThreadIds = new Set(flattenThreads(incoming).map((thread) => thread.id))
-  const localInProgressThreads = flattenThreads(previous).filter(
-    (thread) => inProgressById[thread.id] === true && !incomingThreadIds.has(thread.id),
-  )
-
-  if (localInProgressThreads.length === 0) {
-    return incoming
-  }
-
-  const incomingByProjectName = new Map(incoming.map((group) => [group.projectName, group]))
-  const merged: UiProjectGroup[] = incoming.map((group) => ({
-    projectName: group.projectName,
-    threads: [...group.threads],
-  }))
-
-  for (const thread of localInProgressThreads) {
-    const existingGroup = incomingByProjectName.get(thread.projectName)
-    if (existingGroup) {
-      const mergedGroupIndex = merged.findIndex((group) => group.projectName === thread.projectName)
-      if (mergedGroupIndex >= 0) {
-        merged[mergedGroupIndex] = {
-          projectName: merged[mergedGroupIndex].projectName,
-          threads: [thread, ...merged[mergedGroupIndex].threads],
-        }
-      }
-      continue
-    }
-
-    merged.push({
-      projectName: thread.projectName,
-      threads: [thread],
-    })
-  }
-
-  return merged
-}
-
-function toProjectNameFromWorkspaceRoot(value: string): string {
-  return toProjectName(value)
-}
-
-function getRemoteProjectHostLabel(hostId: string): string {
-  const normalized = hostId.trim()
-  if (!normalized) return ''
-  const separatorIndex = normalized.lastIndexOf(':')
-  return separatorIndex >= 0 ? normalized.slice(separatorIndex + 1) : normalized
-}
-
-function getRemoteProjectDisplayName(remoteProject: NonNullable<WorkspaceRootsState['remoteProjects']>[number]): string {
-  const label = remoteProject.label || toProjectName(remoteProject.remotePath) || remoteProject.id
-  const hostLabel = getRemoteProjectHostLabel(remoteProject.hostId)
-  return hostLabel ? `${label} ${hostLabel}` : label
-}
-
-function getRemoteProjectById(rootsState: WorkspaceRootsState | null): Map<string, NonNullable<WorkspaceRootsState['remoteProjects']>[number]> {
-  const remoteProjects = rootsState?.remoteProjects ?? []
-  return new Map(remoteProjects.map((project) => [project.id, project]))
-}
-
-function getWorkspaceProjectOrderPaths(rootsState: WorkspaceRootsState | null): string[] {
-  if (!rootsState) return []
-  const savedRoots = new Set(rootsState.order)
-  const remoteProjectIds = new Set((rootsState.remoteProjects ?? []).map((project) => project.id))
-  const orderedRoots = rootsState.projectOrder.filter((item) => savedRoots.has(item) || remoteProjectIds.has(item))
-  for (const rootPath of rootsState.order) {
-    if (!orderedRoots.includes(rootPath)) orderedRoots.push(rootPath)
-  }
-  for (const remoteProjectId of remoteProjectIds) {
-    if (!orderedRoots.includes(remoteProjectId)) orderedRoots.push(remoteProjectId)
-  }
-  return orderedRoots
-}
-
-function getWorkspaceProjectOrderNames(
-  rootsState: WorkspaceRootsState | null,
-  duplicateLeafNames: Set<string>,
-): string[] {
-  const remoteProjectsById = getRemoteProjectById(rootsState)
-  return getWorkspaceProjectOrderPaths(rootsState).map((rootPath) => {
-    if (remoteProjectsById.has(rootPath)) return rootPath
-    const normalizedRootPath = normalizePathForUi(rootPath).trim()
-    const leafName = toProjectNameFromWorkspaceRoot(normalizedRootPath)
-    return duplicateLeafNames.has(leafName) ? normalizedRootPath : leafName
-  })
-}
-
-function matchesWorkspaceRootProject(rootPath: string, projectName: string): boolean {
-  const normalizedRootPath = normalizePathForUi(rootPath).trim()
-  return normalizedRootPath === projectName || toProjectNameFromWorkspaceRoot(rootPath) === projectName
-}
-
-export function collectWorkspaceRootPathsForProjectRemoval(
-  rootsState: WorkspaceRootsState,
-  projectName: string,
-): Set<string> {
-  const removedRootPaths = new Set<string>()
-  for (const rootPath of rootsState.order) {
-    if (matchesWorkspaceRootProject(rootPath, projectName)) {
-      removedRootPaths.add(rootPath)
-    }
-  }
-  for (const rootPath of rootsState.active) {
-    if (matchesWorkspaceRootProject(rootPath, projectName)) {
-      removedRootPaths.add(rootPath)
-    }
-  }
-  for (const rootPath of Object.keys(rootsState.labels)) {
-    if (matchesWorkspaceRootProject(rootPath, projectName)) {
-      removedRootPaths.add(rootPath)
-    }
-  }
-  return removedRootPaths
-}
-
-export function buildWorkspaceRootsProjectOrderState(
-  rootsState: WorkspaceRootsState,
-  orderedProjectNames: string[],
-  groups: UiProjectGroup[],
-): Pick<WorkspaceRootsState, 'order' | 'active' | 'projectOrder'> {
-  const remoteProjectIds = new Set((rootsState.remoteProjects ?? []).map((project) => project.id))
-  const rootByProjectName = new Map<string, string>()
-  for (const rootPath of rootsState.order) {
-    const projectName = toProjectNameFromWorkspaceRoot(rootPath)
-    if (!rootByProjectName.has(projectName)) {
-      rootByProjectName.set(projectName, rootPath)
-    }
-  }
-  for (const group of groups) {
-    const cwd = group.threads[0]?.cwd?.trim() ?? ''
-    if (!cwd) continue
-    rootByProjectName.set(group.projectName, cwd)
-  }
-
-  const nextProjectOrder: string[] = []
-  const pushProjectOrderItem = (item: string): void => {
-    if (item && !nextProjectOrder.includes(item)) {
-      nextProjectOrder.push(item)
-    }
-  }
-
-  for (const projectName of orderedProjectNames) {
-    if (remoteProjectIds.has(projectName)) {
-      pushProjectOrderItem(projectName)
-      continue
-    }
-    const rootPath = rootByProjectName.get(projectName)
-    if (rootPath) {
-      pushProjectOrderItem(rootPath)
-    }
-  }
-  for (const item of getWorkspaceProjectOrderPaths(rootsState)) {
-    pushProjectOrderItem(item)
-  }
-
-  const nextOrder = nextProjectOrder.filter((item) => rootsState.order.includes(item))
-  for (const rootPath of rootsState.order) {
-    if (!nextOrder.includes(rootPath)) {
-      nextOrder.push(rootPath)
-    }
-  }
-
-  const nextActive = rootsState.active.filter((rootPath) => nextOrder.includes(rootPath))
-  if (nextActive.length === 0 && nextOrder.length > 0) {
-    nextActive.push(nextOrder[0])
-  }
-
-  return {
-    order: nextOrder,
-    active: nextActive,
-    projectOrder: nextProjectOrder,
-  }
-}
-
-function orderGroupsByWorkspaceProjectOrder(
-  groups: UiProjectGroup[],
-  rootsState: WorkspaceRootsState | null,
-  duplicateLeafNames: Set<string>,
-): UiProjectGroup[] {
-  const order = getWorkspaceProjectOrderNames(rootsState, duplicateLeafNames)
-  if (order.length === 0) return groups
-  const orderIndexByName = new Map(order.map((name, index) => [name, index]))
-  return [...groups].sort((first, second) => {
-    if (isProjectlessGroup(first) || isProjectlessGroup(second)) return 0
-    const firstIndex = orderIndexByName.get(first.projectName) ?? Number.POSITIVE_INFINITY
-    const secondIndex = orderIndexByName.get(second.projectName) ?? Number.POSITIVE_INFINITY
-    if (firstIndex === secondIndex) return 0
-    return firstIndex - secondIndex
-  })
-}
-
-function collectDuplicateProjectLeafNames(groups: UiProjectGroup[], rootsState: WorkspaceRootsState | null): Set<string> {
-  const rootByLeafName = new Map<string, Set<string>>()
-  const canonicalWorkspaceRootCountsByLeafName = new Map<string, number>()
-  const addPath = (value: string): void => {
-    const normalizedPath = normalizePathForUi(value).trim()
-    if (!normalizedPath) return
-    const leafName = toProjectName(normalizedPath)
-    const existing = rootByLeafName.get(leafName) ?? new Set<string>()
-    existing.add(normalizedPath)
-    rootByLeafName.set(leafName, existing)
-  }
-
-  for (const rootPath of rootsState?.order ?? []) {
-    const normalizedRootPath = normalizePathForUi(rootPath).trim()
-    if (!normalizedRootPath) continue
-    const leafName = toProjectName(normalizedRootPath)
-    if (!isManagedCodexWorktreePath(normalizedRootPath)) {
-      canonicalWorkspaceRootCountsByLeafName.set(leafName, (canonicalWorkspaceRootCountsByLeafName.get(leafName) ?? 0) + 1)
-    }
-    addPath(rootPath)
-  }
-  for (const group of groups) {
-    for (const thread of group.threads) {
-      const normalizedCwd = normalizePathForUi(thread.cwd).trim()
-      const leafName = toProjectName(normalizedCwd)
-      const isRegisteredRoot = rootsState?.order.some((rootPath) => normalizePathForUi(rootPath).trim() === normalizedCwd) === true
-      if (isManagedCodexWorktreePath(normalizedCwd) && !isRegisteredRoot && canonicalWorkspaceRootCountsByLeafName.get(leafName) === 1) continue
-      addPath(thread.cwd)
-    }
-  }
-
-  const duplicateLeafNames = new Set<string>()
-  for (const [leafName, paths] of rootByLeafName.entries()) {
-    if (paths.size > 1) duplicateLeafNames.add(leafName)
-  }
-  return duplicateLeafNames
-}
-
-function isManagedCodexWorktreePath(value: string): boolean {
-  return value.includes('/.codex/worktrees/')
-}
-
-function disambiguateProjectGroupsByCwd(
-  groups: UiProjectGroup[],
-  rootsState: WorkspaceRootsState | null,
-): UiProjectGroup[] {
-  const duplicateLeafNames = collectDuplicateProjectLeafNames(groups, rootsState)
-  if (duplicateLeafNames.size === 0) return groups
-
-  const uniqueCanonicalWorkspaceRootLeafNames = new Set<string>()
-  const duplicateCanonicalWorkspaceRootLeafNames = new Set<string>()
-  const canonicalWorkspaceRootByLeafName = new Map<string, string>()
-  const registeredWorkspaceRoots = new Set<string>()
-  for (const rootPath of rootsState?.order ?? []) {
-    const normalizedRootPath = normalizePathForUi(rootPath).trim()
-    if (!normalizedRootPath) continue
-    registeredWorkspaceRoots.add(normalizedRootPath)
-    if (isManagedCodexWorktreePath(normalizedRootPath)) continue
-    const leafName = toProjectName(normalizedRootPath)
-    if (uniqueCanonicalWorkspaceRootLeafNames.has(leafName)) {
-      uniqueCanonicalWorkspaceRootLeafNames.delete(leafName)
-      duplicateCanonicalWorkspaceRootLeafNames.add(leafName)
-      canonicalWorkspaceRootByLeafName.delete(leafName)
-    } else if (!duplicateCanonicalWorkspaceRootLeafNames.has(leafName)) {
-      uniqueCanonicalWorkspaceRootLeafNames.add(leafName)
-      canonicalWorkspaceRootByLeafName.set(leafName, normalizedRootPath)
-    }
-  }
-
-  const disambiguatedGroups: UiProjectGroup[] = []
-  const groupsByProjectName = new Map<string, UiProjectGroup>()
-  for (const group of groups) {
-    for (const thread of group.threads) {
-      const normalizedCwd = normalizePathForUi(thread.cwd).trim()
-      const leafName = toProjectName(normalizedCwd)
-      const isRegisteredRoot = registeredWorkspaceRoots.has(normalizedCwd)
-      const isCanonicalWorktreeThread = isManagedCodexWorktreePath(normalizedCwd)
-        && !isRegisteredRoot
-        && uniqueCanonicalWorkspaceRootLeafNames.has(leafName)
-      let projectName = group.projectName
-      if (isCanonicalWorktreeThread && duplicateLeafNames.has(leafName)) {
-        projectName = canonicalWorkspaceRootByLeafName.get(leafName) ?? group.projectName
-      } else if (normalizedCwd && duplicateLeafNames.has(leafName)) {
-        projectName = normalizedCwd
-      }
-      const nextThread = thread.projectName === projectName ? thread : { ...thread, projectName }
-      const existingGroup = groupsByProjectName.get(projectName)
-      if (existingGroup) {
-        existingGroup.threads.push(nextThread)
-      } else {
-        const nextGroup = { projectName, threads: [nextThread] }
-        groupsByProjectName.set(projectName, nextGroup)
-        disambiguatedGroups.push(nextGroup)
-      }
-    }
-  }
-
-  return disambiguatedGroups
-}
-
-function addWorkspaceRootPlaceholderGroups(
-  groups: UiProjectGroup[],
-  rootsState: WorkspaceRootsState | null,
-  duplicateLeafNames: Set<string>,
-): UiProjectGroup[] {
-  if (!rootsState || (rootsState.order.length === 0 && (rootsState.remoteProjects ?? []).length === 0)) return groups
-  const existingProjectNames = new Set(groups.map((group) => group.projectName))
-  const nextGroups = [...groups]
-  const remoteProjectsById = getRemoteProjectById(rootsState)
-
-  for (const rootPath of getWorkspaceProjectOrderPaths(rootsState)) {
-    if (remoteProjectsById.has(rootPath)) {
-      if (existingProjectNames.has(rootPath)) continue
-      nextGroups.push({ projectName: rootPath, threads: [] })
-      existingProjectNames.add(rootPath)
-      continue
-    }
-    const normalizedRootPath = normalizePathForUi(rootPath).trim()
-    if (!normalizedRootPath) continue
-    const leafName = toProjectNameFromWorkspaceRoot(normalizedRootPath)
-    const projectName = duplicateLeafNames.has(leafName) ? normalizedRootPath : leafName
-    if (existingProjectNames.has(projectName)) continue
-    nextGroups.push({ projectName, threads: [] })
-    existingProjectNames.add(projectName)
-  }
-
-  return nextGroups
-}
-
-function toOptimisticThreadTitle(message: string): string {
-  const firstLine = message
-    .split('\n')
-    .map((line) => line.trim())
-    .find((line) => line.length > 0)
-
-  if (!firstLine) return 'Untitled thread'
-  return firstLine.slice(0, 80)
-}
-
-function toForkedThreadTitle(title: string): string {
-  const normalizedTitle = title.trim() || 'Untitled thread'
-  return /^fork:\s+/iu.test(normalizedTitle) ? normalizedTitle : `Fork: ${normalizedTitle}`
-}
-
-function isProjectlessGroup(group: UiProjectGroup): boolean {
-  return group.threads.some((thread) => thread.cwd.trim().length === 0 || isProjectlessChatPath(thread.cwd))
-}
-
-export function filterGroupsByWorkspaceRoots(
-  groups: UiProjectGroup[],
-  rootsState: WorkspaceRootsState | null,
-): UiProjectGroup[] {
-  const duplicateLeafNames = collectDuplicateProjectLeafNames(groups, rootsState)
-  const disambiguatedGroups = disambiguateProjectGroupsByCwd(groups, rootsState)
-  const groupsWithWorkspaceRoots = addWorkspaceRootPlaceholderGroups(disambiguatedGroups, rootsState, duplicateLeafNames)
-  if (!rootsState || (rootsState.order.length === 0 && (rootsState.remoteProjects ?? []).length === 0)) return groupsWithWorkspaceRoots
-  const allowedProjectNames = new Set<string>()
-  for (const projectName of getWorkspaceProjectOrderNames(rootsState, duplicateLeafNames)) {
-    allowedProjectNames.add(projectName)
-  }
-  const filteredGroups = groupsWithWorkspaceRoots.filter((group) => allowedProjectNames.has(group.projectName) || isProjectlessGroup(group))
-  return orderGroupsByWorkspaceProjectOrder(filteredGroups, rootsState, duplicateLeafNames)
-}
+import {
+  WORKED_MESSAGE_TYPE,
+  BACKGROUND_THREAD_PAGINATION_DELAY_MS,
+  EVENT_SYNC_DEBOUNCE_MS,
+  GLOBAL_SERVER_REQUEST_SCOPE,
+  MODEL_FALLBACK_ID,
+  RATE_LIMIT_REFRESH_DEBOUNCE_MS,
+  REASONING_EFFORT_OPTIONS,
+  RECENT_THREAD_MESSAGE_LOAD_REUSE_MS,
+  TURN_START_FOLLOW_UP_SYNC_DELAY_MS,
+  addWorkspaceRootPlaceholderGroups,
+  areGroupArraysEqual,
+  areMessageArraysEqual,
+  areStringArraysEqual,
+  areThreadArraysEqual,
+  areThreadFieldsEqual,
+  areTurnActivitiesEqual,
+  areTurnSummariesEqual,
+  buildWorkspaceRootsProjectOrderState,
+  collectWorkspaceRootPathsForProjectRemoval,
+  collectDuplicateProjectLeafNames,
+  disambiguateProjectGroupsByCwd,
+  filterGroupsByWorkspaceRoots,
+  findAdjacentThreadId,
+  flattenThreads,
+  formatTurnDuration,
+  getRemoteProjectDisplayName,
+  getWorkspaceProjectOrderNames,
+  getWorkspaceProjectOrderPaths,
+  insertTurnSummaryMessage,
+  isProjectlessGroup,
+  isUnsupportedChatGptModelError,
+  matchesWorkspaceRootProject,
+  mergeIncomingWithLocalInProgressThreads,
+  mergeMessages,
+  mergeProjectOrder,
+  normalizeMessageText,
+  mergeThreadGroups,
+  omitKey,
+  omitKeys,
+  orderGroupsByProjectOrder,
+  orderGroupsByWorkspaceProjectOrder,
+  parseIsoTimestamp,
+  pruneThreadStateMap,
+  removePersistedLiveMessages,
+  removeRedundantLiveAgentMessages,
+  removeThreadFromGroups,
+  reorderStringArray,
+  toForkedThreadTitle,
+  toOptimisticThreadTitle,
+  toProjectNameFromWorkspaceRoot,
+  upsertMessage,
+  type TurnActivityState,
+  type TurnCompletedInfo,
+  type TurnErrorState,
+  type TurnStartedInfo,
+  type TurnSummaryState,
+} from './desktopStateThreadHelpers'
+export {
+  buildWorkspaceRootsProjectOrderState,
+  collectWorkspaceRootPathsForProjectRemoval,
+  filterGroupsByWorkspaceRoots,
+  findAdjacentThreadId,
+  removeThreadFromGroups,
+} from './desktopStateThreadHelpers'
 
 export function useDesktopState() {
   const projectGroups = ref<UiProjectGroup[]>([])
